requirements.txt
asgiref==3.3.1
boto3==1.17.68
botocore==1.20.68
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-cors-headers==3.5.0
django-debug-toolbar==2.2
django-filter==2.4.0
django-multiselectfield==0.1.12
django-sqlprint-middleware==0.1.3
django-storages==1.11.1
djangorestframework==3.12.2
fcm-django==0.3.10
idna==2.10
jmespath==0.10.0
numpy==1.19.5
pandas==1.1.5
psycopg2-binary==2.8.6
pyfcm==1.4.9
Pygments==2.8.1
pygments-pprint-sql==0.1.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-http-client==3.3.2
pytz==2021.1
requests==2.25.1
s3transfer==0.4.2
sendgrid==6.7.0
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
stripe==2.56.0
twilio==6.57.0
urllib3==1.26.4
XlsxWriter==1.4.4
